+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JYCG5WBR6W96AMXC0ZH08YCC"
+        "01JYEBTEGR10GMFBM81AQT7KWW"
     )
     r"""The id of the resource"""