+++ packages/orq-rc/RELEASES.md
@@ -1789,3 +1789,13 @@ Based on:
 - [python v3.9.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.0] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.0 - packages/orq-rc
+
+## 2025-06-23 12:23:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.0 (2.632.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.1] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.1 - packages/orq-rc